Allow hubmap/main.js to be used without Browserify
Wrap the HubMap functionality in something close to a UMD which supports
Node style require and raw use in the browser. This allows us to make
use of Browserify for the main embed builds but potentially allows the
HubMap functionaility to be used in a project that doesn't use
Browserify.